requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-auth
beautifulsoup4
requests
aiogram
python-dotenv

